refactor: remove barrel exports, use direct imports
Delete all index.ts barrel files and update imports to use direct paths.
Improves tree-shaking, build performance, and eliminates circular dep risk.
