--- lib/validation/index.ts
@@ -1,18 +1,0 @@
-// Type exports
-export type { ValidationResult, Validator } from './types';
-
-// Validator function exports
-export {
-  validateRequired,
-  validateEmail,
-  validateUrl,
-  validateMinLength,
-  validateMaxLength,
-  validatePattern,
-  validateMatch,
-  validateOptional,
-  validateRange,
-  validateDigitsOnly,
-  validateExactLength,
-  composeValidators,
-} from './validators';
