--- ui/common/buttons/index.ts
@@ -1,21 +1,0 @@
-/**
- * Button components module
- *
- * Provides consistent button styling and behavior across the application.
- */
-
-export { default as Button } from "./Button";
-export type { ButtonProps } from "./Button";
-
-export { default as CopyButton } from "./CopyButton";
-export type { CopyButtonProps } from "./CopyButton";
-
-export { default as IconButton } from "./IconButton";
-export type { IconButtonProps } from "./IconButton";
-
-// Style utilities
-export {
-  INLINE_SELECTOR_TRIGGER_CLASSES,
-  OUTLINE_ACTION_BUTTON_CLASSES,
-  INLINE_ACTION_BUTTON_CLASSES,
-} from "./styles";
