--- ui/common/feedback/index.ts
@@ -1,14 +1,0 @@
-/**
- * Feedback components module
- *
- * Provides components for user feedback, status indicators, and loading states.
- */
-
-export { default as Badge } from "./Badge";
-export type { BadgeProps } from "./Badge";
-
-export { default as Spinner } from "./Spinner";
-export type { SpinnerProps } from "./Spinner";
-
-export { default as Alert } from "./Alert";
-export type { AlertProps } from "./Alert";
