--- ui/common/layout/index.ts
@@ -1,14 +1,0 @@
-/**
- * Layout components module
- *
- * Provides container and structural components for consistent page layouts.
- */
-
-export { default as Card } from "./Card";
-export type { CardProps } from "./Card";
-
-export { default as Section } from "./Section";
-export type { SectionProps } from "./Section";
-
-export { default as Divider } from "./Divider";
-export type { DividerProps } from "./Divider";
