+++ ui/profile/ProfileEditor.tsx
@@ -21,8 +21,9 @@ import { sanitizeUsernameInput } from "@/lib/profile/usernamePolicy";
 import useVerificationFlow from "@/ui/social/useVerificationFlow";
 import { useEditsStore, type ParsedLink, type FormState } from "@/ui/profile/store";
 import type { Profile, EnrichedProfileLink } from "@/lib/profile/types";
-import { Alert, Button } from "@/ui/common";
-import { withFieldBorderState } from "@/ui/common/forms";
+import Alert from "@/ui/common/feedback/Alert";
+import Button from "@/ui/common/buttons/Button";
+import { withFieldBorderState } from "@/ui/common/forms/styles";
 
 const FIELD_CLASS =
   `w-full rounded-2xl border px-3 py-2 text-sm bg-transparent outline-hidden text-gray-800 placeholder-gray-400 ${withFieldBorderState("border-[#0a1126]/60")}`;
