+++ ui/signup/SocialLinkInput.tsx
@@ -3,8 +3,8 @@ import { normalizeSocialUsername, buildSocialUrl } from "@/lib/profile/usernameN
 import type { SocialPlatform } from "@/lib/profile/usernameNormalizer";
 import { isValidUrl } from "@/lib/validation/validators";
 import HelpIcon from "@/ui/common/HelpIcon";
-import { Alert } from "@/ui/common";
-import { withFieldBorderState } from "@/ui/common/forms";
+import Alert from "@/ui/common/feedback/Alert";
+import { withFieldBorderState } from "@/ui/common/forms/styles";
 
 const PLATFORM_OPTIONS = [
   { key: "X", label: "X (Twitter)" },
