--- ui/thread/index.ts
@@ -1,8 +1,0 @@
-export { ThreadCard } from './ThreadCard';
-export { ThreadFeed } from './ThreadFeed';
-export { ThreadComposer } from './ThreadComposer';
-export { BoardHeader } from './BoardHeader';
-export { SidebarNav } from './SidebarNav';
-export { CreateBoardModal } from './CreateBoardModal';
-export { BoardSelector } from './BoardSelector';
-export { ThreadBoard } from './ThreadBoard';
