+++ ui/verification/ProfileVerification.tsx
@@ -10,7 +10,7 @@ import { buildZcashUri } from "@/lib/zcash/zcashUtils";
 import useVerificationPolling from "@/ui/verification/useVerificationPolling";
 import ProgressStep from "@/ui/verification/ProgressStep";
 import { useMessagingStore } from "@/ui/verification/store";
-import { Alert } from "@/ui/common";
+import Alert from "@/ui/common/feedback/Alert";
 
 const SIGNIN_ADDR = "u1lff6xhc9p2c3aefrms5624aqd5mdlys87xcu0u0g3rynnjfs4g5nf0u5q8sczex3jctc2xesauktvdr9gd77zauaejje3zrdpj4uppssdmzzu33lfkzc9y0hlq7rt94kt4rqpq6d4h8a0px597htclme3pav3wft4k94u4pqqn3h4dmdp8wcvvumgqak5ynwy7qm6e797t356ud38we";
